+++ R/layout.R
@@ -1,8 +1,8 @@
-layout_ragged_rows <- function(x, free = list(), transpose = FALSE) {
+layout_ragged_rows <- function(x, free = list(), align = "start", transpose = FALSE) {
     r <- vctrs::vec_group_rle(x)
     n <- vctrs::field(r, "length")
     i <- rep.int(seq_along(n), n)
-    j <- sequence(n)
+    j <- sequence(n, from = if (align == "end") max(n) - n + 1L else 1L)
 
     if (transpose) {
       layout <- list(PANEL = vctrs::vec_seq_along(x), ROW = j, COL = i)
@@ -19,6 +19,6 @@ layout_ragged_rows <- function(x, free = list(), transpose = FALSE) {
     layout
 }
 
-layout_ragged_cols <- function(x, free = list(), transpose = FALSE) {
-  layout_ragged_rows(x, free = free, transpose = !transpose)
+layout_ragged_cols <- function(x, free = list(), align = "start", transpose = FALSE) {
+  layout_ragged_rows(x, free = free, align = align, transpose = !transpose)
 }