+++ tests/testthat/test-layout.R
@@ -25,3 +25,10 @@ test_that("free scales work", {
   expect_equal(cols_free_y$SCALE_X, c(1, 1, 1))
   expect_equal(cols_free_y$SCALE_Y, c(1, 2, 3))
 })
+
+test_that("panels can be aligned to the end", {
+  expect_equal(layout_ragged_rows(c("A", "B", "B"), align = "start")$COL, c(1, 1, 2))
+  expect_equal(layout_ragged_cols(c("A", "B", "B"), align = "start")$ROW, c(1, 1, 2))
+  expect_equal(layout_ragged_rows(c("A", "B", "B"), align = "end")$COL, c(2, 1, 2))
+  expect_equal(layout_ragged_cols(c("A", "B", "B"), align = "end")$ROW, c(2, 1, 2))
+})